requirements.txt
configparser==3.5.0
Twisted==16.6.0
GitPython==2.1.1